refactor : 도메인 모듈 리패키징
