--- shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/MemberNewsRepository.kt
@@ -1,7 +1,0 @@
-package com.mashup.shorts.domain.member
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface MemberNewsRepository : JpaRepository<MemberNews, Long>
