+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/membercategory/MemberCategoryRepository.kt
@@ -1,4 +1,4 @@
-package com.mashup.shorts.domain.member
+package com.mashup.shorts.domain.member.membercategory
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
