+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/member/membernws/MemberNews.kt
@@ -1,21 +1,22 @@
-package com.mashup.shorts.domain.member
+package com.mashup.shorts.domain.member.membernws
 
 import com.mashup.shorts.domain.BaseEntity
-import com.mashup.shorts.domain.newscard.NewsCard
+import com.mashup.shorts.domain.member.Member
+import com.mashup.shorts.domain.news.News
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 
-@Table(name = "member_news_card")
+@Table(name = "member_news")
 @Entity
-class MemberNewsCard(
+class MemberNews(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     val member: Member,
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "news_card_id")
-    val newsCard: NewsCard,
+    @JoinColumn(name = "news_id")
+    val news: News,
 ) : BaseEntity()
